- Move handling of multiple schemas per search cluster down into SearchCluster. - Prepare for multiple schemas per search cluster for streaming too.
